backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
langchain
langchain-groq
langgraph
pydantic
httpx
